# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface Creator {
  id: string;
  username: string;
  displayName: string;
  bio: string;
  avatarUrl?: string | null;
  /** Solana public key (base58) */
  walletAddress: string;
  createdAt: string;
}

export interface CreateCreatorBody {
  /**
   * @minLength 3
   * @maxLength 30
   * @pattern ^[a-z0-9_]+$
   */
  username: string;
  /**
   * @minLength 1
   * @maxLength 60
   */
  displayName: string;
  /** @maxLength 280 */
  bio: string;
  avatarUrl?: string | null;
  /** Solana public key (base58, 32-44 chars) */
  walletAddress: string;
}

export interface Tip {
  id: string;
  creatorUsername: string;
  /** Solana transaction signature */
  signature: string;
  /** Amount in lamports (1 SOL = 1,000,000,000 lamports) */
  amountLamports: number;
  tipperName: string;
  tipperWallet?: string | null;
  message: string;
  createdAt: string;
}

export type TipWithCreator = Tip & {
  creatorDisplayName: string;
  creatorAvatarUrl?: string | null;
};

export interface RecordTipBody {
  signature: string;
  /** @minimum 1 */
  amountLamports: number;
  /** @maxLength 60 */
  tipperName: string;
  tipperWallet?: string | null;
  /** @maxLength 280 */
  message: string;
}

export interface CreatorStats {
  totalTips: number;
  totalLamports: number;
  uniqueTippers: number;
  biggestTipLamports: number;
}

export interface PlatformStats {
  totalCreators: number;
  totalTips: number;
  totalLamports: number;
}

export type CreatorWithStats = Creator & {
  totalTips: number;
  totalLamports: number;
};

export type ListCreatorsParams = {
  search?: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type ListCreatorTipsParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type ListRecentTipsParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetTopCreatorsParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};