# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * List creators, optionally filtered by search term, ordered by most recent activity.
 * @summary List creators
 */
export const listCreatorsQueryLimitDefault = 20;
export const listCreatorsQueryLimitMax = 100;

export const ListCreatorsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  limit: zod.coerce
    .number()
    .min(1)
    .max(listCreatorsQueryLimitMax)
    .default(listCreatorsQueryLimitDefault),
});

export const ListCreatorsResponseItem = zod.object({
  id: zod.string(),
  username: zod.string(),
  displayName: zod.string(),
  bio: zod.string(),
  avatarUrl: zod.string().nullish(),
  walletAddress: zod.string().describe("Solana public key (base58)"),
  createdAt: zod.coerce.date(),
});
export const ListCreatorsResponse = zod.array(ListCreatorsResponseItem);

/**
 * @summary Create a new creator profile
 */
export const createCreatorBodyUsernameMin = 3;
export const createCreatorBodyUsernameMax = 30;

export const createCreatorBodyUsernameRegExp = new RegExp("^[a-z0-9_]+$");
export const createCreatorBodyDisplayNameMax = 60;

export const createCreatorBodyBioMax = 280;

export const CreateCreatorBody = zod.object({
  username: zod
    .string()
    .min(createCreatorBodyUsernameMin)
    .max(createCreatorBodyUsernameMax)
    .regex(createCreatorBodyUsernameRegExp),
  displayName: zod.string().min(1).max(createCreatorBodyDisplayNameMax),
  bio: zod.string().max(createCreatorBodyBioMax),
  avatarUrl: zod.string().nullish(),
  walletAddress: zod
    .string()
    .describe("Solana public key (base58, 32-44 chars)"),
});

/**
 * @summary Get a creator by username
 */
export const GetCreatorParams = zod.object({
  username: zod.coerce.string(),
});

export const GetCreatorResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  displayName: zod.string(),
  bio: zod.string(),
  avatarUrl: zod.string().nullish(),
  walletAddress: zod.string().describe("Solana public key (base58)"),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List tips received by a creator
 */
export const ListCreatorTipsParams = zod.object({
  username: zod.coerce.string(),
});

export const listCreatorTipsQueryLimitDefault = 50;
export const listCreatorTipsQueryLimitMax = 100;

export const ListCreatorTipsQueryParams = zod.object({
  limit: zod.coerce
    .number()
    .min(1)
    .max(listCreatorTipsQueryLimitMax)
    .default(listCreatorTipsQueryLimitDefault),
});

export const ListCreatorTipsResponseItem = zod.object({
  id: zod.string(),
  creatorUsername: zod.string(),
  signature: zod.string().describe("Solana transaction signature"),
  amountLamports: zod
    .number()
    .describe("Amount in lamports (1 SOL = 1,000,000,000 lamports)"),
  tipperName: zod.string(),
  tipperWallet: zod.string().nullish(),
  message: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListCreatorTipsResponse = zod.array(ListCreatorTipsResponseItem);

/**
 * Record a tip transaction after the client has sent SOL on Solana Devnet. The server verifies the signature exists on-chain.
 * @summary Record a tip after on-chain confirmation
 */
export const RecordTipParams = zod.object({
  username: zod.coerce.string(),
});

export const recordTipBodyTipperNameMax = 60;

export const recordTipBodyMessageMax = 280;

export const RecordTipBody = zod.object({
  signature: zod.string(),
  amountLamports: zod.number().min(1),
  tipperName: zod.string().max(recordTipBodyTipperNameMax),
  tipperWallet: zod.string().nullish(),
  message: zod.string().max(recordTipBodyMessageMax),
});

/**
 * @summary Get aggregate stats for a creator
 */
export const GetCreatorStatsParams = zod.object({
  username: zod.coerce.string(),
});

export const GetCreatorStatsResponse = zod.object({
  totalTips: zod.number(),
  totalLamports: zod.number(),
  uniqueTippers: zod.number(),
  biggestTipLamports: zod.number(),
});

/**
 * @summary Recent tips across all creators (global tip wall)
 */
export const listRecentTipsQueryLimitDefault = 30;
export const listRecentTipsQueryLimitMax = 100;

export const ListRecentTipsQueryParams = zod.object({
  limit: zod.coerce
    .number()
    .min(1)
    .max(listRecentTipsQueryLimitMax)
    .default(listRecentTipsQueryLimitDefault),
});

export const ListRecentTipsResponseItem = zod
  .object({
    id: zod.string(),
    creatorUsername: zod.string(),
    signature: zod.string().describe("Solana transaction signature"),
    amountLamports: zod
      .number()
      .describe("Amount in lamports (1 SOL = 1,000,000,000 lamports)"),
    tipperName: zod.string(),
    tipperWallet: zod.string().nullish(),
    message: zod.string(),
    createdAt: zod.coerce.date(),
  })
  .and(
    zod.object({
      creatorDisplayName: zod.string(),
      creatorAvatarUrl: zod.string().nullish(),
    }),
  );
export const ListRecentTipsResponse = zod.array(ListRecentTipsResponseItem);

/**
 * @summary Get global platform stats
 */
export const GetPlatformStatsResponse = zod.object({
  totalCreators: zod.number(),
  totalTips: zod.number(),
  totalLamports: zod.number(),
});

/**
 * @summary Top creators by total tips received
 */
export const getTopCreatorsQueryLimitDefault = 10;
export const getTopCreatorsQueryLimitMax = 50;

export const GetTopCreatorsQueryParams = zod.object({
  limit: zod.coerce
    .number()
    .min(1)
    .max(getTopCreatorsQueryLimitMax)
    .default(getTopCreatorsQueryLimitDefault),
});

export const GetTopCreatorsResponseItem = zod
  .object({
    id: zod.string(),
    username: zod.string(),
    displayName: zod.string(),
    bio: zod.string(),
    avatarUrl: zod.string().nullish(),
    walletAddress: zod.string().describe("Solana public key (base58)"),
    createdAt: zod.coerce.date(),
  })
  .and(
    zod.object({
      totalTips: zod.number(),
      totalLamports: zod.number(),
    }),
  );
export const GetTopCreatorsResponse = zod.array(GetTopCreatorsResponseItem);