# lib/api-zod/src/generated/types/createCreatorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCreatorBody {
  /**
   * @minLength 3
   * @maxLength 30
   * @pattern ^[a-z0-9_]+$
   */
  username: string;
  /**
   * @minLength 1
   * @maxLength 60
   */
  displayName: string;
  /** @maxLength 280 */
  bio: string;
  avatarUrl?: string | null;
  /** Solana public key (base58, 32-44 chars) */
  walletAddress: string;
}