# lib/api-zod/src/generated/types/creator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */

export interface Creator {
  id: string;
  username: string;
  displayName: string;
  bio: string;
  avatarUrl?: string | null;
  /** Solana public key (base58) */
  walletAddress: string;
  createdAt: Date;
}