# lib/api-zod/src/generated/types/creatorWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */
import type { Creator } from "./creator";

export type CreatorWithStats = Creator & {
  totalTips: number;
  totalLamports: number;
};