# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalCreators: number;
  totalTips: number;
  totalLamports: number;
}