# lib/api-zod/src/generated/types/recordTipBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */

export interface RecordTipBody {
  signature: string;
  /** @minimum 1 */
  amountLamports: number;
  /** @maxLength 60 */
  tipperName: string;
  tipperWallet?: string | null;
  /** @maxLength 280 */
  message: string;
}