# lib/api-zod/src/generated/types/tip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */

export interface Tip {
  id: string;
  creatorUsername: string;
  /** Solana transaction signature */
  signature: string;
  /** Amount in lamports (1 SOL = 1,000,000,000 lamports) */
  amountLamports: number;
  tipperName: string;
  tipperWallet?: string | null;
  message: string;
  createdAt: Date;
}