# lib/api-zod/src/generated/types/tipWithCreator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spray — Solana creator tip jar API
 * OpenAPI spec version: 0.1.0
 */
import type { Tip } from "./tip";

export type TipWithCreator = Tip & {
  creatorDisplayName: string;
  creatorAvatarUrl?: string | null;
};